+++ gwbackupy/tests/test_helpers.py
@@ -12,6 +12,7 @@ from gwbackupy.helpers import (
     json_load,
     parse_date,
     is_rate_limit_exceeded,
+    random_string,
 )
 
 
@@ -88,3 +89,8 @@ def test_is_rate_limit_exceeded():
     data2[0]["error"]["details"].append(dict())
     e = HttpError(Resp(403, "Forbidden"), json.dumps(data2).encode("utf8"))
     assert not is_rate_limit_exceeded(e)
+
+
+def test_random_string():
+    for i in range(32):
+        assert len(random_string(i)) == i