+++ CHANGELOG.md
@@ -2,10 +2,17 @@
 
 ## DEV
 
+
+
+## 0.7.0
+
+- Enh #38: kill signals handling
+
+## 0.6.0
+
 - Enh: namespace refactoring `gwb` -> `gwbackupy`
 - Enh #22: restore missing Gmail messages (`--restore-missing`)
 - Bug: fixed date filters parsing
-- Enh #38: kill signals handling
 
 ## 0.5.0